[TASK] Move plugins from model folder to module root folder according to best practice from devdocs
