+++ Plugin/Agreements/AfterGetContent.php
@@ -3,7 +3,7 @@
  * Copyright © 2015 FireGento e.V. - All rights reserved.
  * See LICENSE.md bundled with this module for license details.
  */
-namespace FireGento\MageSetup\Model\Plugin;
+namespace FireGento\MageSetup\Plugin\Agreements;
 
 use Magento\Cms\Model\Template\Filter as TemplateFilter;
