+++ Plugin/Pricing/AroundRenderPlugin.php
@@ -3,7 +3,7 @@
  * Copyright © 2015 FireGento e.V. - All rights reserved.
  * See LICENSE.md bundled with this module for license details.
  */
-namespace FireGento\MageSetup\Model\Plugin;
+namespace FireGento\MageSetup\Plugin\Pricing;
 
 /**
  * Class AroundRenderPlugin
